Renamed services to repository Split services and interface into separate files.
