+++ CleanArchitectureBlazor/CleanArchitectureBlazor/Components/Pages/Dashboard.razor
@@ -2,9 +2,9 @@
 @using Domain
 @using CleanArchitectureBlazor.Services
 @inject IEventService EventService
-@inject IShiftService ShiftService
-@inject IStaffService StaffService
-@inject IStaffAssignmentService StaffAssignmentService
+@inject IShiftRepository ShiftService
+@inject IStaffRepository StaffService
+@inject IStaffAssignmentRepository StaffAssignmentService
 @rendermode InteractiveServer
 
 <PageTitle>Dashboard - Medical First Aid Manager</PageTitle>
