+++ CleanArchitectureBlazor/CleanArchitectureBlazor/Components/Pages/Events.razor
@@ -2,7 +2,7 @@
 @using Domain
 @using CleanArchitectureBlazor.Services
 @inject IEventService EventService
-@inject IShiftService ShiftService
+@inject IShiftRepository ShiftService
 @inject NavigationManager Navigation
 @rendermode InteractiveServer
