+++ CleanArchitectureBlazor/CleanArchitectureBlazor/Components/Pages/Shifts.razor
@@ -1,7 +1,7 @@
 @page "/shifts"
 @using Domain
 @using CleanArchitectureBlazor.Services
-@inject IShiftService ShiftService
+@inject IShiftRepository ShiftService
 @inject IEventService EventService
 @inject NavigationManager Navigation
 @rendermode InteractiveServer
