+++ CleanArchitectureBlazor/CleanArchitectureBlazor/Components/Pages/Staff.razor
@@ -1,7 +1,7 @@
 @page "/staff"
 @using CleanArchitectureBlazor.Services
 
-@inject IStaffService StaffService
+@inject IStaffRepository StaffService
 @inject NavigationManager Navigation
 @rendermode InteractiveServer
