+++ CleanArchitectureBlazor/CleanArchitectureBlazor/Services/StaffRepository.cs
@@ -4,26 +4,11 @@ using Microsoft.EntityFrameworkCore;
 
 namespace CleanArchitectureBlazor.Services;
 
-/// <summary>
-/// Service for managing staff
-/// </summary>
-public interface IStaffService
-{
-    Task<List<Staff>> GetAllStaffAsync();
-    Task<Staff?> GetStaffByIdAsync(int id);
-    Task<Staff> CreateStaffAsync(Domain.Staff staff);
-    Task<Staff> UpdateStaffAsync(Domain.Staff staff);
-    Task DeleteStaffAsync(int id);
-    Task<List<Domain.Staff>> GetActiveStaffAsync();
-    Task<List<Domain.Staff>> GetStaffByRoleAsync(StaffRole role);
-    Task<bool> IsEmailUniqueAsync(string email, int? excludeId = null);
-}
-
-public class StaffService : IStaffService
+public class StaffRepository : IStaffRepository
 {
     private readonly ApplicationDbContext _context;
 
-    public StaffService(ApplicationDbContext context)
+    public StaffRepository(ApplicationDbContext context)
     {
         _context = context;
     }
